放置処置をLeftAloneControllerに委譲
